+++ src/guifuncs.h
@@ -39,6 +39,6 @@ void freeList(struct List *, int);
 void showAvatarImage(STRPTR, STRPTR, Object *, Object *, struct RenderHook *);
 ULONG renderfunct(struct RenderHook *, Object *, struct gpRender *);
 void workOnUpdate(void);
-void switchRightSidebar(ULONG);
+void switchPage(ULONG, Object *);
 
 #endif